Refactor content layout pages phase 1
