+++ src/Files.App/Data/Parameters/ColumnParam.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT License. See the LICENSE.
 
 using Files.App.Views;
-using Files.App.Views.LayoutModes;
+using Files.App.Views.ContentLayouts;
 using Microsoft.UI.Xaml.Controls;
 
 namespace Files.App.Data.Parameters
