+++ src/Files.App/GlobalUsings.cs
@@ -48,7 +48,7 @@ global using global::Files.App.UserControls;
 global using global::Files.App.ViewModels;
 global using global::Files.App.ViewModels.UserControls;
 global using global::Files.App.Views;
-global using global::Files.App.Views.LayoutModes;
+global using global::Files.App.Views.ContentLayouts;
 global using global::Files.App.Views.Shells;
 
 // Files.Core
