+++ src/Files.App/Views/ContentLayouts/BaseLayoutPage.cs
@@ -4,7 +4,7 @@
 using CommunityToolkit.WinUI.UI;
 using Files.App.Helpers.ContextFlyouts;
 using Files.App.UserControls.Menus;
-using Files.App.ViewModels.LayoutModes;
+using Files.App.ViewModels.ContentLayouts;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Controls.Primitives;
@@ -26,12 +26,12 @@ using DispatcherQueueTimer = Microsoft.UI.Dispatching.DispatcherQueueTimer;
 using SortDirection = Files.Core.Data.Enums.SortDirection;
 using VanaraWindowsShell = Vanara.Windows.Shell;
 
-namespace Files.App.Views.LayoutModes
+namespace Files.App.Views.ContentLayouts
 {
 	/// <summary>
 	/// Represents the base class which every layout page must derive from
 	/// </summary>
-	public abstract class BaseLayout : Page, IBaseLayout, INotifyPropertyChanged
+	public abstract class BaseLayoutPage : Page, IBaseLayoutPage, INotifyPropertyChanged
 	{
 		private readonly DispatcherQueueTimer jumpTimer;
 
@@ -253,7 +253,7 @@ namespace Files.App.Views.LayoutModes
 
 		protected abstract ItemsControl ItemsControl { get; }
 
-		public BaseLayout()
+		public BaseLayoutPage()
 		{
 			PreviewPaneViewModel = Ioc.Default.GetRequiredService<PreviewPaneViewModel>();
 			ItemManipulationModel = new ItemManipulationModel();
