+++ src/Files.App/Views/ContentLayouts/ColumnViewBase.xaml
@@ -1,6 +1,6 @@
 ﻿<!--  Copyright (c) 2023 Files Community. Licensed under the MIT License. See the LICENSE.  -->
-<local:StandardViewBase
-	x:Class="Files.App.Views.LayoutModes.ColumnViewBase"
+<local:BaseGroupableLayoutPage
+	x:Class="Files.App.Views.ContentLayouts.ColumnViewBase"
 	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	xmlns:cloud="using:Files.App.Utils.Cloud"
@@ -8,7 +8,7 @@
 	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
 	xmlns:filesystem="using:Files.App.Utils"
 	xmlns:helpers="using:Files.App.Helpers"
-	xmlns:local="using:Files.App.Views.LayoutModes"
+	xmlns:local="using:Files.App.Views.ContentLayouts"
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	xmlns:storage="using:Files.App.Utils.Storage"
 	xmlns:uc="using:Files.App.UserControls"
@@ -17,7 +17,7 @@
 	x:Name="PageRoot"
 	mc:Ignorable="d">
 
-	<local:StandardViewBase.Resources>
+	<local:BaseGroupableLayoutPage.Resources>
 
 		<converters:StringToBrushConverter x:Key="StringToBrushConverter" />
 		<wctconverters:BoolNegationConverter x:Key="BoolNegationConverter" />
@@ -123,7 +123,7 @@
 			</Setter>
 		</Style>
 
-	</local:StandardViewBase.Resources>
+	</local:BaseGroupableLayoutPage.Resources>
 
 	<Grid x:Name="RootGrid" ContextFlyout="{x:Bind BaseContextMenuFlyout}">
 
@@ -481,4 +481,4 @@
 		</Canvas>
 
 	</Grid>
-</local:StandardViewBase>
+</local:BaseGroupableLayoutPage>
