+++ src/Files.App/Views/ContentLayouts/ColumnViewBrowser.xaml
@@ -1,13 +1,13 @@
 <!--  Copyright (c) 2023 Files Community. Licensed under the MIT License. See the LICENSE.  -->
-<local:BaseLayout
-	x:Class="Files.App.Views.LayoutModes.ColumnViewBrowser"
+<local:BaseLayoutPage
+	x:Class="Files.App.Views.ContentLayouts.ColumnViewBrowser"
 	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	xmlns:controls="using:CommunityToolkit.WinUI.UI.Controls"
 	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
 	xmlns:i="using:Microsoft.Xaml.Interactivity"
 	xmlns:icore="using:Microsoft.Xaml.Interactions.Core"
-	xmlns:local="using:Files.App.Views.LayoutModes"
+	xmlns:local="using:Files.App.Views.ContentLayouts"
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	x:Name="PageRoot"
 	mc:Ignorable="d">
@@ -38,4 +38,4 @@
 		</controls:BladeItem>
 	</controls:BladeView>
 
-</local:BaseLayout>
+</local:BaseLayoutPage>
