+++ src/Files.App/Views/ContentLayouts/ColumnViewBrowser.xaml.cs
@@ -3,7 +3,7 @@
 
 using CommunityToolkit.WinUI.UI;
 using CommunityToolkit.WinUI.UI.Controls;
-using Files.App.ViewModels.LayoutModes;
+using Files.App.ViewModels.ContentLayouts;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Navigation;
@@ -11,12 +11,12 @@ using Windows.Storage;
 using static Files.App.Constants;
 using static Files.App.Helpers.PathNormalization;
 
-namespace Files.App.Views.LayoutModes
+namespace Files.App.Views.ContentLayouts
 {
 	/// <summary>
 	/// Represents the browser page of Column View
 	/// </summary>
-	public sealed partial class ColumnViewBrowser : BaseLayout
+	public sealed partial class ColumnViewBrowser : BaseLayoutPage
 	{
 		protected override uint IconSize => Browser.ColumnViewBrowser.ColumnViewSizeSmall;
