+++ src/Files.App/Views/ContentLayouts/DetailsLayoutBrowser.xaml
@@ -1,6 +1,6 @@
 <!--  Copyright (c) 2023 Files Community. Licensed under the MIT License. See the LICENSE.  -->
-<local:StandardViewBase
-	x:Class="Files.App.Views.LayoutModes.DetailsLayoutBrowser"
+<local:BaseGroupableLayoutPage
+	x:Class="Files.App.Views.ContentLayouts.DetailsLayoutBrowser"
 	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	xmlns:behaviors="using:Files.App.Data.Behaviors"
@@ -11,7 +11,7 @@
 	xmlns:helpers="using:Files.App.Helpers"
 	xmlns:i="using:Microsoft.Xaml.Interactivity"
 	xmlns:icore="using:Microsoft.Xaml.Interactions.Core"
-	xmlns:local="using:Files.App.Views.LayoutModes"
+	xmlns:local="using:Files.App.Views.ContentLayouts"
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	xmlns:storage="using:Files.App.Utils.Storage"
 	xmlns:toolkit="using:CommunityToolkit.WinUI.UI.Controls"
@@ -31,13 +31,16 @@
 		</icore:EventTriggerBehavior>
 	</i:Interaction.Behaviors>
 
-	<local:StandardViewBase.Resources>
+	<local:BaseGroupableLayoutPage.Resources>
 		<ResourceDictionary>
 			<ResourceDictionary.MergedDictionaries>
 				<ResourceDictionary Source="ms-appx:///ResourceDictionaries/DefaultGridSplitterStyle.xaml" />
 			</ResourceDictionary.MergedDictionaries>
 
+			<converters:StringToBrushConverter x:Key="StringToBrushConverter" />
+			<converters:BoolToSelectionModeConverter x:Key="BoolToSelectionModeConverter" />
 			<wctconverters:BoolNegationConverter x:Key="BoolNegationConverter" />
+			<wctconverters:BoolToVisibilityConverter x:Key="BoolToVisibilityConverter" />
 			<wctconverters:BoolToVisibilityConverter
 				x:Key="NegatedBoolToVisibilityConverter"
 				FalseValue="Visible"
@@ -46,9 +49,6 @@
 				x:Key="EmptyObjectToObjectConverter"
 				EmptyValue="Collapsed"
 				NotEmptyValue="Visible" />
-			<converters:StringToBrushConverter x:Key="StringToBrushConverter" />
-			<converters:BoolToSelectionModeConverter x:Key="BoolToSelectionModeConverter" />
-			<wctconverters:BoolToVisibilityConverter x:Key="BoolToVisibilityConverter" />
 
 			<Style
 				x:Name="HeaderGridSplitterStyle"
@@ -166,7 +166,7 @@
 			</Style>
 
 		</ResourceDictionary>
-	</local:StandardViewBase.Resources>
+	</local:BaseGroupableLayoutPage.Resources>
 
 	<Grid
 		x:Name="RootGrid"
@@ -1531,4 +1531,4 @@
 		</Canvas>
 
 	</Grid>
-</local:StandardViewBase>
+</local:BaseGroupableLayoutPage>
