+++ src/Files.App/Views/ContentLayouts/DetailsLayoutBrowser.xaml.cs
@@ -16,12 +16,12 @@ using Windows.System;
 using Windows.UI.Core;
 using SortDirection = Files.Core.Data.Enums.SortDirection;
 
-namespace Files.App.Views.LayoutModes
+namespace Files.App.Views.ContentLayouts
 {
 	/// <summary>
 	/// Represents the browser page of Details View
 	/// </summary>
-	public sealed partial class DetailsLayoutBrowser : StandardViewBase
+	public sealed partial class DetailsLayoutBrowser : BaseGroupableLayoutPage
 	{
 		private const int TAG_TEXT_BLOCK = 1;
