+++ src/Files.App/Views/ContentLayouts/GridViewBrowser.xaml
@@ -1,6 +1,6 @@
 <!--  Copyright (c) 2023 Files Community. Licensed under the MIT License. See the LICENSE.  -->
-<local:StandardViewBase
-	x:Class="Files.App.Views.LayoutModes.GridViewBrowser"
+<local:BaseGroupableLayoutPage
+	x:Class="Files.App.Views.ContentLayouts.GridViewBrowser"
 	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	xmlns:cloud="using:Files.App.Utils.Cloud"
@@ -10,7 +10,7 @@
 	xmlns:helpers="using:Files.App.Helpers"
 	xmlns:i="using:Microsoft.Xaml.Interactivity"
 	xmlns:icore="using:Microsoft.Xaml.Interactions.Core"
-	xmlns:local="using:Files.App.Views.LayoutModes"
+	xmlns:local="using:Files.App.Views.ContentLayouts"
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	xmlns:storage="using:Files.App.Utils.Storage"
 	xmlns:toolkit="using:CommunityToolkit.WinUI.UI.Controls"
@@ -31,8 +31,9 @@
 		</icore:EventTriggerBehavior>
 	</i:Interaction.Behaviors>
 
-	<local:StandardViewBase.Resources>
+	<local:BaseGroupableLayoutPage.Resources>
 
+		<converters:BoolToSelectionModeConverter x:Key="BoolToSelectionModeConverter" />
 		<converters:StringToBrushConverter x:Key="StringToBrushConverter" />
 		<wctconverters:BoolNegationConverter x:Key="BoolNegationConverter" />
 		<wctconverters:BoolToVisibilityConverter
@@ -43,7 +44,6 @@
 			x:Key="EmptyObjectToObjectConverter"
 			EmptyValue="Collapsed"
 			NotEmptyValue="Visible" />
-
 		<wctconverters:BoolToObjectConverter
 			x:Key="NegatedBoolToOpacityConverter"
 			FalseValue="1"
@@ -52,7 +52,6 @@
 			x:Key="BoolToVisibilityConverter"
 			FalseValue="Collapsed"
 			TrueValue="Visible" />
-		<converters:BoolToSelectionModeConverter x:Key="BoolToSelectionModeConverter" />
 
 		<DataTemplate x:Name="GridViewBrowserTemplate" x:DataType="filesystem:ListedItem">
 			<UserControl>
@@ -528,7 +527,7 @@
 				</Setter.Value>
 			</Setter>
 		</Style>
-	</local:StandardViewBase.Resources>
+	</local:BaseGroupableLayoutPage.Resources>
 
 	<Grid x:Name="RootGrid" ContextFlyout="{x:Bind BaseContextMenuFlyout}">
 
@@ -774,4 +773,4 @@
 		</Canvas>
 
 	</Grid>
-</local:StandardViewBase>
+</local:BaseGroupableLayoutPage>
