+++ src/Files.App/Views/ContentLayouts/IBaseLayoutPage.cs
@@ -1,12 +1,12 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using Files.App.ViewModels.LayoutModes;
+using Files.App.ViewModels.ContentLayouts;
 using Microsoft.UI.Xaml.Controls;
 
-namespace Files.App.Views.LayoutModes
+namespace Files.App.Views.ContentLayouts
 {
-	public interface IBaseLayout : IDisposable
+	public interface IBaseLayoutPage : IDisposable
 	{
 		bool IsRenamingItem { get; }
